requirements.txt
# Основные библиотеки
python-telegram-bot==20.3
llama-cpp-python>=0.2.0
langchain>=0.0.200
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pymupdf>=1.22.0
transformers>=4.30.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
torch>=2.0.0

# Служебные библиотеки
tqdm>=4.66.0
psutil>=5.9.5
numpy>=1.24.0
jsonlines>=3.1.0
uvicorn>=0.22.0
fastapi>=0.100.0

# Обработка данных
pandas>=2.0.0
pyyaml>=6.0.0
pillow>=10.0.0
